JI:883501 BBM-39328 jsmn: add function to minify json
This commit adds jsmn_compact() which takes a pre-parsed token list
that indexes the parsed string and re-serialises the data in a
compact form (ie, without unnecessary whitespace).

NB: this commit is a rewrite by df of an earlier version by sttang,
maintaining the same interface.  This version removes the self-
implemented stack and instead uses recursion, while fixing a
formatting bug.

Change-Id: I5ab119d9a93c85cc9ac05132f079bd16f7016d3e