JI:883501 BBM-39328 jsmn: avoid nullptr dereference in parser
The jsmn parser may be called with a null token array pointer
to count the number of tokens that need to be allocated.  Static
analysis shows that if the parser state is not correctly reset,
a null dereference will occur when the parser is validating
that the parse tree balances.

This commit prevents this condition by returning before the
final json correctness check (which requires a token list).