JI:883501 BBM-39328 jsmn: Add a utility function to fix escaped quotes in json strings
Description:
To embed json elements into a json string, one would have to escape
the double quotes for the parser to properly parse the string. However,
to parse the embedded json elements these escaped quotes must be fixed
to be valid.

Testing Done:
Updated unit test of conf_mgr to test this case and verified passing
the result.

Change-Id: I012d39f5cebd2feb5bcc1787715382a83a3e014b